requirements.txt
astroid==2.4.2
control==0.9.0
cycler==0.10.0
descartes==1.1.0
ffmpeg-python==0.2.0
future==0.18.2
gitdb==4.0.5
GitPython==3.1.7
install==1.3.4
isort==5.4.2
joblib==0.16.0
JPype1==1.3.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.3.0
mccabe==0.6.1
mizani==0.7.1
numpy==1.19.0
packaging==20.4
palettable==3.3.0
pandas==1.0.5
patsy==0.5.1
Pillow==7.2.0
plotnine==0.7.0
git+https://github.com/oist/ecsu-pyevolver@0.4.3
git+https://gitlab.com/oist-ecsu/statsmeasures.git@0.1.0
pygame==2.0.1
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-ranges==0.2.0
pytictoc==1.5.1
pytz==2020.1
scikit-learn==0.24.2
scipy==1.5.1
seaborn==0.11.2
semver==2.10.2
six==1.15.0
sklearn==0.0
smmap==3.0.4
statsmodels==0.11.1
threadpoolctl==2.2.0
toml==0.10.1
tqdm==4.62.1
typed-ast==1.4.2
typing-extensions==3.10.0.2
version-query==1.1.0
wrapt==1.12.1
